infra/requirements.txt
aws_cdk.core==1.134.0
aws_cdk.aws_ec2==1.134.0
aws_cdk.aws_elasticloadbalancing==1.134.0
aws_cdk.aws_elasticloadbalancingv2==1.134.0
aws_cdk.aws_elasticloadbalancingv2-targets==1.134.0
aws_cdk.aws_route53==1.134.0
aws_cdk.aws_route53_targets==1.134.0
aws_cdk.aws_lambda==1.134.0
aws_cdk.aws_iam==1.134.0
aws_cdk.aws_sns==1.134.0
aws_cdk.aws_cloudwatch==1.134.0
aws_cdk.aws_events==1.134.0
aws_cdk.aws_events_targets==1.134.0
aws_cdk.aws_logs==1.134.0
pylint==2.8.2
pytest

